requirements.txt
asgiref==3.8.1
gunicorn
Django
django-cors-headers==4.3.1
djangorestframework==3.15.1
mysqlclient==2.2.4
psycopg2-binary==2.9.9
setuptools==69.5.1
sqlparse==0.5.0
wheel==0.43.0
python-dotenv